changed type of Bar_trades (#715)

+++ ta4j-core/src/main/java/org/ta4j/core/Bar.java
@@ -67,7 +67,7 @@ public interface Bar extends Serializable {
     /**
      * @return the number of trades in the period
      */
-    int getTrades();
+    long getTrades();
 
     /**
      * @return the whole traded amount of the period
